requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
numpy==2.2.4
opencv-python==4.11.0.86
requests==2.32.3
urllib3==2.4.0
python-dotenv==1.1.0